feat: remove version checker thread in onhost
