+++ agent-control/src/agent_type/runtime_config/version_config.rs
@@ -1,9 +1,6 @@
 use serde::Deserialize;
 
-use crate::agent_type::{
-    definition::Variables, error::AgentTypeError, templates::Templateable,
-    version_config::VersionCheckerInterval,
-};
+use crate::agent_type::{definition::Variables, error::AgentTypeError, templates::Templateable};
 
 /// Represents the configuration for version checks.
 #[derive(Debug, Deserialize, Clone, PartialEq)]
@@ -17,10 +14,6 @@ pub struct OnHostVersionConfig {
     /// If not provided, the entire output will be used.
     #[serde(default)]
     pub(crate) regex: Option<String>,
-
-    /// The duration to wait between version checks.
-    #[serde(default)]
-    pub(crate) interval: VersionCheckerInterval,
 }
 
 impl Templateable for OnHostVersionConfig {
@@ -28,7 +21,6 @@ impl Templateable for OnHostVersionConfig {
         Ok(Self {
             command: self.command.template_with(variables)?,
             regex: self.regex,
-            interval: self.interval,
         })
     }
 }
